rewrite mocks and snapshots
