--- packages/webdriverio/tests/__mocks__/fs.js
@@ -1,4 +1,0 @@
-const fs = jest.genMockFromModule('fs')
-fs.writeFileSync = jest.fn()
-fs.existsSync = jest.fn(() => true)
-module.exports = fs
