+++ packages/webdriverio/tests/__mocks__/repl.ts
@@ -1,0 +1,9 @@
+class ReplMock {
+    start = jest.fn().mockReturnValue({ on: this.on.bind(this) })
+
+    on (event: string, cb: Function) {
+        cb(event)
+    }
+}
+
+export default new ReplMock()
